+++ rust/src/api/yande_client.rs
@@ -2,11 +2,11 @@ use flutter_rust_bridge::DartFnFuture;
 use crate::yande::model::Post;
 use crate::yande::YANDE_CLIENT;
 
-pub async fn get_posts(tags: Vec<String>, limit: usize, page: usize) -> anyhow::Result<Vec<Post>> {
+pub async fn get_posts(tags: Vec<String>, limit: u32, page: u32) -> anyhow::Result<Vec<Post>> {
     YANDE_CLIENT.get_posts(tags, limit, page).await
 }
 
-pub async fn get_similar(post_id: u64) -> anyhow::Result<crate::yande::model::Similar> {
+pub async fn get_similar(post_id: i64) -> anyhow::Result<crate::yande::model::Similar> {
     YANDE_CLIENT.get_similar(post_id).await
 }
